refactor(desktop): split oversized files in BE and FE
Split FormatDialog.tsx (874 LOC) into 7 files:
- utils/formatUtils.ts, utils/tableHelpers.ts, utils/formatColumns.tsx
- FormatDialogHeader.tsx, FormatTableSection.tsx, DownloadOptionsPanel.tsx
- FormatDialog.tsx reduced to 325 LOC

Split formats.rs (513 LOC) into formats/ module:
- formats/models.rs (109 LOC) for struct definitions
- formats/mod.rs (413 LOC) for commands and tests

All 45 Rust tests pass. TypeScript compiles cleanly.
