+++ crates/rdlp-desktop/src-tauri/src/commands/formats/mod.rs
@@ -7,79 +7,14 @@
 //! serialisable [`FormatListResponse`] containing format details,
 //! subtitles, thumbnail URL, and duration.
 
-use serde::{Deserialize, Serialize};
+mod models;
+
+pub(crate) use models::*;
+
 use tauri::State;
 
 use crate::error::AppError;
 use crate::state::AppState;
-
-/// Frontend-facing format information.
-///
-/// A simplified projection of [`rdlp_types::Format`] that exposes
-/// only the fields the UI needs, without internal download URLs.
-#[derive(Debug, Serialize)]
-pub struct FormatInfo {
-    /// Unique format identifier (e.g. "137", "hls-720").
-    pub format_id: String,
-    /// File extension (e.g. "mp4", "webm").
-    pub ext: String,
-    /// Human-readable quality label (e.g. "720p", "1080p").
-    pub format_note: Option<String>,
-    /// Video width in pixels.
-    pub width: Option<u32>,
-    /// Video height in pixels.
-    pub height: Option<u32>,
-    /// Frames per second.
-    pub fps: Option<f64>,
-    /// Total bitrate in kbps (video + audio).
-    pub tbr: Option<f64>,
-    /// Video codec name (e.g. "h264", "vp9").
-    pub vcodec: Option<String>,
-    /// Audio codec name (e.g. "aac", "opus").
-    pub acodec: Option<String>,
-    /// File size in bytes (exact or approximate).
-    pub filesize: Option<u64>,
-    /// Video bitrate in kbps.
-    pub vbr: Option<f64>,
-    /// Audio bitrate in kbps.
-    pub abr: Option<f64>,
-    /// Audio sampling rate in Hz.
-    pub asr: Option<u32>,
-    /// Download protocol (e.g. "https", "m3u8_native").
-    pub protocol: String,
-    /// Whether this format contains a video stream.
-    pub has_video: bool,
-    /// Whether this format contains an audio stream.
-    pub has_audio: bool,
-}
-
-/// Subtitle availability for a single language.
-#[derive(Debug, Serialize)]
-pub struct SubtitleInfo {
-    /// Language code (e.g. "en", "ja").
-    pub lang: String,
-    /// Subtitle format extensions (kept as strings because extractors
-    /// may return unknown formats that do not map to `SubtitleFormat`).
-    pub formats: Vec<String>,
-}
-
-/// Complete response for the `get_formats` command.
-///
-/// Contains all metadata the frontend needs to render the format
-/// selection UI.
-#[derive(Debug, Serialize)]
-pub struct FormatListResponse {
-    /// Video title.
-    pub title: String,
-    /// Available download formats.
-    pub formats: Vec<FormatInfo>,
-    /// Available subtitle languages and their formats.
-    pub subtitles: Vec<SubtitleInfo>,
-    /// URL of the best thumbnail image.
-    pub thumbnail_url: Option<String>,
-    /// Video duration in seconds.
-    pub duration: Option<f64>,
-}
 
 /// Retrieve available formats for a URL.
 ///
@@ -173,41 +108,6 @@ pub async fn get_formats(
         thumbnail_url: info.thumbnail,
         duration: info.duration,
     })
-}
-
-/// Format metadata sent from the frontend for expression validation.
-///
-/// Mirrors the fields from [`FormatInfo`] that the format selector
-/// uses for filtering and ranking, allowing expressions like
-/// `bv[height<=1080]+ba` to match correctly.
-#[derive(Debug, Deserialize)]
-pub struct FormatData {
-    /// Unique format identifier.
-    pub(crate) format_id: String,
-    /// File extension (e.g. "mp4", "webm").
-    pub(crate) ext: String,
-    /// Video width in pixels.
-    pub(crate) width: Option<u32>,
-    /// Video height in pixels.
-    pub(crate) height: Option<u32>,
-    /// Frames per second.
-    pub(crate) fps: Option<f64>,
-    /// Total bitrate in kbps.
-    pub(crate) tbr: Option<f64>,
-    /// Video codec name.
-    pub(crate) vcodec: Option<String>,
-    /// Audio codec name.
-    pub(crate) acodec: Option<String>,
-    /// File size in bytes.
-    pub(crate) filesize: Option<u64>,
-    /// Video bitrate in kbps.
-    pub(crate) vbr: Option<f64>,
-    /// Audio bitrate in kbps.
-    pub(crate) abr: Option<f64>,
-    /// Audio sampling rate in Hz.
-    pub(crate) asr: Option<u32>,
-    /// Download protocol string (e.g. "https", "m3u8_native").
-    pub(crate) protocol: String,
 }
 
 /// Validate a yt-dlp format expression and return the matching format IDs.
